Adds Integrations and Integration Instances
You can now fetch a list of integrations from the GraphQL backend, and
then create instances of those integrations on your User Account.

Then you can enable/disable the integration instance on a workspace, or
enable/disable a workspace from the integration instance itself.

The database will need to be wiped before this code is run for the first
time, if you've been iterating on any version of this branch.