Adds Pulumi script for deploying GraphQL API
This creates a Pulumi script for deploying our GraphQL API. It includes
a tiny EC2 instance to run the service itself, auto-deployment from a
habitat package that is created automatically, and an SSL load balancer.

Eventually, we probably want to have it create a database as well. But
we live to fight another day, yes?