feat(si-graphql,si-web-app): Add components
This adds components to the system. Components are "created" when an
integration is added to the system; so for example, the AWS integration
would create Server, Network, etc. components.

This commit includes a generic interface for querying for lists of
components, or specific components - this should make it (relatively)
easy to create new Components.

The Web UI now includes a data table showing the components that are
available in a given workspace, and on a given integration. This is an
early example of how we can use the same data to populate the nodes
themselves when we get there.

Next up is adding prototypes for all the different types of Components;
this commit only includes the 'Server' component.

Various refactorings and types are also included in this commit. Of
particular note is the shift to having all the graphql modules included
in a single global `app.modules`, which reflects the current best
practice in graphql modules land.