feat: Add SSH Key components and entities
This adds an SSH Key component and entities. Also has the first
example of an actual "agent" style backend, in that it creates
real keys.

This is so we can build up from saying `SshKey.create()` and get back a
completely auto-generated key. This is heading toward us being able to
do a "Service" that runs on a "Server" that gets created on EC2, without
us ever having to specify... anything.

I think it'll be cool.

This is A++ a prototype - we should, um, clearly not actually ship this
code. Proves the point, though.