chore: Docker containers (#36)
* chore: Docker containers

This branch introduces a full container build system for everything. It
consists of a couple multi-stage build containers, one for Rust services
and another for Typescript projects.

It's all driven through the toplevel makefile.

It also adds test container support - if you set the environment
variable TEST_IN_CONTAINERS to anything, it will automatically use
the baseline development container to build and test your code. That
container should be build every day, ensuring everything is always
up to date and cached.

It also adds a daily rebuild of the base build and test image, which should speed up
the build process significantly, along with a workflow that pushes new releases whenever
a commit is pushed to master.