bug: Fixing addresses and no more squash
In fixing the no-squash bug, I forgot to fix the fact that we now create
a bullshit docker container with a bunch of stuff in it we do not need,
due to layers.

This fixes that error, by making one long-ass bullshit line in the base
dockerfile image.

Also, we switch to IPv4, to save on our own sanity.