feat: Integration instances
Add Integration Instances to the system. These are what get created when
you create an Integration. They can also be enabled/disabled on a per
Organization or per Workspace basis.

This commit also adds in `has_list` and `in_list`, which let you go
between an array of id's on an object to the objects themselves (via the
List endpoint for the object type), and the reverse (find all the
objects that match the current objects ID inside an array).

It also adds 'LIKE' and 'NOT LIKE' support to the query AST, which lets
you do `%` and `_` based globbing of strings.