feat(*): Create SSH Key Entities (#57)
Ultimately, this commit adds a fully working SSH Key component, entity, actor framework, and a front-end that streams output. It's pretty cool. You'll dig it.

This commit re-adds migration functionality, now based on a manually
incremented "version" integer in the objects you wish to migrate. It
also adds Integrations, along with a `listIntegrations` query.

Add Integration Instances to the system. These are what get created when
you create an Integration. They can also be enabled/disabled on a per
Organization or per Workspace basis.

This commit also adds in `has_list` and `in_list`, which let you go
between an array of id's on an object to the objects themselves (via the
List endpoint for the object type), and the reverse (find all the
objects that match the current objects ID inside an array).

It also adds 'LIKE' and 'NOT LIKE' support to the query AST, which lets
you do `%` and `_` based globbing of strings.

This adds vernemq to the stack.

You now have an Agent with 3 parts:

* Client
* Server
* Finalizer

They all work, and just run in the service. Probably they should be
their own binaries, since that would let them scale horizontally. But,
hey, it works!

Also, you can stream the entity events for a workspace back from
graphql. It is spooky magic, friends.