feat: Add external API Gateway
This is the first pass at our external API gateway. It wraps external
APIs in a GRPC layer, understands our authentication and authorizatio
model, knows how to deal with keys and credentials, and wires together
the API calls. It's going to be neat.

It contains both Typescript and a Rust generated client library for the
service itself, which will be consumed by whoever needs to call the
service.