feat: external api gateway (#61)
* feat: add external API Gateway

This is the first pass at our external API gateway. It wraps external
APIs in a GRPC layer andwires together the API calls. It's going to be neat.

It contains both Typescript and a Rust generated client library for the
service itself, which will be consumed by whoever needs to call the
service.

* feat: SSH Key Agent now understands AWS

This allows the SSH Key Agent to create AWS keys, and ensures that a
bogus AWS integration is enabled on every new account that gets created.
It then uses the external API gateway to create a new SSH keypair in AWS, and
store the entity locally.