chore(framework): Add the component/entity/actor framework
This commit creates the SI CEA framework. It lets us build the core
domain logic of our systems (the Component/Entity/Actor parts) easily,
and centralizes as much of the logic as we can.

Currently the only component using the framework is SSH - all other
components are going to break, due to changes in the implementation of
some shared traits.

Long live cea!