feat: fully create an EKS cluster with nodegroup
This change adds full endpoint and certificate information which allows
a user to use the resulting Kubernetes cluster as per usual with
`kubectl` and other tooling.

This change also refactors a good deal of the
`si-aws-eks-cluster-runtime` and `si-external-api-gateway` code to
incrementally improve data validation and have better common semantics
across services, programming languages, and protobuf semantics.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>