feat: AWS EKS Runtime, Web Registry, CEA, and K8S
* feat: si-aws-eks-cluster-runtime

The Web UI is now generated from a registry of SI Components, each of
which specifies its own GraphQL queries. This mirrors our internal
domain model, and should make things much easier to extend. Eventually,
we can auto-generate these, just like we do the GraphQL layer.

It also refactors the (very) hacky components we threw together, so that
they encapsulate their state correctly.

* feat: generate the menu, fix pagination

This commit makes the menu on the left auto generated, fixes pagination,
and tweaks the theme just a bit (so that "darker" isn't the highlight
color).

It also switches EntityList to javascript rather than typescript,
because the vue-apollo typings are just not correct.

* feat: Syncing of EC2 SSH Key resources (and global ones)

This adds the sync feature, although minus a button to trigger the
event.

* feat: sync button for ssh keys

* feat: support all request/response data when creating EKS cluster

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* feat: Streaming Entity Events

You can now stream entity events generically, and use the tenancy as
as a filtering mechanism.Yay!

* feat: capture EKS status on cluster create

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* feat: support sync EKS cluster entities

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* feat: add EKS Nodegroup to a Cluster

Yeah, it's great!

Note that we're slamming this idea right into the EKS Cluster Entity for
the moment so this isn't purely ideal but we need to see this work
first.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* chore: add support for watching multiple paths via `make watch`

This change helps to make the `make watch` task more responsive to
changes in protobuf files in Rust projects.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* feat: fully create an EKS cluster with nodegroup

This change adds full endpoint and certificate information which allows
a user to use the resulting Kubernetes cluster as per usual with
`kubectl` and other tooling.

This change also refactors a good deal of the
`si-aws-eks-cluster-runtime` and `si-external-api-gateway` code to
incrementally improve data validation and have better common semantics
across services, programming languages, and protobuf semantics.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* chore(framework): Add the component/entity/actor framework

This commit creates the SI CEA framework. It lets us build the core
domain logic of our systems (the Component/Entity/Actor parts) easily,
and centralizes as much of the logic as we can.

Currently the only component using the framework is SSH - all other
components are going to break, due to changes in the implementation of
some shared traits.

Long live cea!

* chore(si-graphql-api): Fix security vulns

Fixes a pile of security vulnerabilities!

* chore(si-external-api-gateway): Fix vulns

Fix the ridiculous acorn/mkdirp vulnerabilities.

* chore(webui): Fix security vulns

This upgrades some of our NPM packages. It's very exciting.

* chore: deleting deprecated code

It's time for Bugbear to say goodbye. I have it on my hard drive, and it
lives forever in the history if we really need it.

Goodbye!

* chore: Fix tests

This updates the si-data tests to reflect the changes in the Storable
trait.

* chore: Fix type hints, remove console logs

This makes the `make test` task work cleanly

* chore: Deploy to k8s

This commit deploys the software to k8s. It includes the required
deployment YAML, although it cannot contain all of it (secrets!
controllers! random rbac!).

It also has a few commits to do things like configure mqtt correctly
from a configuration file.

It also moves the containers from Github's package repository, which
would not work with AWS, to the AWS ECR.

Long live our k8s overlord.

Co-authored-by: Fletcher Nichol <fnichol@nichol.ca>