chore: Refactor the registry
This refactors the registry to no longer use a massive god object in
`component.ts`. Instead, it uses the `PropObject` as the root of most
things, of which Components, Entitys, etc. are only derivations.

This sets us up to add features like relationships between objects, and
greatly simplifies code generation.

As of this commit, only protobuf code generation is working.