Code generate core models, components, and entities (#71)
* WIP: strting si-kubernetes refactor

* Re-add the default type attrs

* feat: Auto-generate rust code for components

This adds the MVP of adding rust code for components. It creates a 'gen'
sudirectory, which has a subdirectory per component, and writes out the
component.rs file.

* chore WIP: refactor si-kubernetes using protobuf codegen

* feat WIP: Use the generated component

This refactors the component model in si-k8s to use the generated model
instead. It also fixes a bug in linked props, which made them use the
underlying properties name and description, rather than the higher level
property name.

* WIP: Fixing up graphql api, and kubernetesVersion enum!

* WIP: Generate the GraphQL API

This commit switches the GraphQL API generator from using the protocol buffers to
using the si-registry components!

* WIP: Fixes to the generator typescript

* WIP si-kubernetes, making entity work

* chore: Refactor the registry

This refactors the registry to no longer use a massive god object in
`component.ts`. Instead, it uses the `PropObject` as the root of most
things, of which Components, Entitys, etc. are only derivations.

This sets us up to add features like relationships between objects, and
greatly simplifies code generation.

As of this commit, only protobuf code generation is working.

* WIP: Refactored the data definitions

* WIP: graphql api works again

Fixed bugs in the protobuf we generated, and now the graphql api works

* WIP agent dispatch

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

* Updates to si-registry

* chore: Account for global object/service naming

* chore: Refactored si-data

Refactor si-data to use the renamed generated assets.

* WIP: halfway to si-account

* WIP: Account works with the GraphQL API

End to end!

* WIP: Works for me

* chore: Authorization, Get, List, Create all work

This fixes authorization, get, list, and create - at least for the auth
service.

* feat: Si-account working, registry understands associations

Login is working, registry and si-graphql both understand associations.

We're back, baby!

* chore: Cleanup si-account

* chore: distributed tracing!

Congratulations, there is some new distributed tracing support!

A wild honeycomb appears!

* chore: Swith to babel/webpack

This switches the si-graphql-api and the si-registry to babel and
webpack, and uses nodemon to run the service.

* Fixing typescript errors

* chore: Switch to webpack/babel

This switches the server to webpack/babel/nodemon, rather than
ts-node-dev, which is broken on Node 14. It also brings us more in line
with the best practices in the javascript ecosystem, such as they are.

* Massive refactoring of si-kubernetes to support codegen

* Refactoring cea Entity/EntityEvent traits with codegen

* Refactor cea/data traits to simplify Rust codegen

* chore: Re-integrate with web ui, fix account build bug

This commit re-integrates with the web ui, and it also fixes a bug in
the generated code that was producing false failures on tenant id
checks.

* chore: switch to item from object

* Initial implementation of `service` for components

* Implement generic Entity::create service

* [codegen] Update registry codegen

* Wire in Kubernetes server, minus AgentServer dispatch

* chore: don't generate a `service.rs` if there are no service methods

* feat: Relationships on properties

You can now delcare relationships on properties. This commit implements
two:

* Updates, which signifies that this property updates another property.
* Either, which signifies that either this property *or* another
  property must be set.

* chore: test fixes

This fixes some test coverage, and brings in a small implementation of
an atomically updated integer (which I may later revert, because it may
not be neccessary after all).

Co-authored-by: Adam Jacob <adam@systeminit.com>