chore: fixing up imported si-web-ui code
This cleans up some of the imported si-web-ui code. It removes three
dependencies that are no longer needed (lodash, deepdash, and
project-name-generator,) re-generates the graphql-types to match the
current API, removes an un-neccesary manual graphql query, and fixes the
types up in the Auth module.

A new "Profile" class is introduced, which stores the user,
BillingAccount, Workspaces, and the default Workspace (currently
whatever the "first" workspace is.)