bug(si-registry): fix race awaiting child process
This fixes a race condition on awaiting a spawned child process. Rather
than `await onExit`, we should set the `onExit` callback first, and then
simply ignore the resolution of the promise. That way we will listen for
the signal from the children, even if they exit (and perhaps get
reaped?) before we get a chance to set the signal handler.

Good times, OS X!