feat: create new entities within a changeset
This commit finishes the changeset create work. If you provide a
changeSetId when you call the create method, we will now populate the
appropriate fields in siStorable, and create the right ID for the new
object.

There is some magic to check out inside of `db.rs` - essentially, if
your storable implementation has the ability to have a changeset id,
then we will atomically update the entry counter inside the changeset.
That is what will let you have N edits to the same object, at different
parts of the changeset. (So you can edit an object, take an action, then
edit the object again, and that will all work.)