feat: change set creation, done right (#104)
* feat: change set creation, done right

This commit makes it so change set creation requires only the
workspaceId. It also extends the model to have a hasMany association
using the `item` interface (and exposed a lingering issue with how
hasMany was implemented via scopeByTenantId).

So you can now create a change set, and then pass the changeSetId on
create to an entity, and the rest is history. To see the list of
entries in a changeSet, you chase the changeSetEntries association.

* chore: remove the superflous changeSetEntry

I thought we would need these, but it turns out - we don't!

Goodbye change set entries, we hardly knew ye.