feat: add si-pedant, our functional testing suite (#108)
* feat: add si-pedant, our functional testing suite

This commit adds si-pedant, a Cypress powered functional testing suite.
It covers both the GraphQL API and the Web UI, and is driven by an
embedded Chrome service. You can run the tests in interactive dev mode
with `npm start`, or run them on the console with `npm test`.

This also fixes a bug in the ChangeSet code, where if you submit an
entry with a changeSetId set to the empty string, we would fail to ever
return from the atomic change tracking system. It was.. bad news. :)

There is also a test to prove we won't regress in this case. Good times!

* chore: clean up debug messages

A few debug messages snuck in. This is a commit of the results of a
clean run of the registry.