bug(si-pedant): Add automatic retrying on custom assertions (#110)
This commit adds automatic retrying on our custom graphql extensions to
Cypress. This means that if, for whatever reason, Couchbase takes a few
seconds to catch up to the change in data, the test will still pass. All
it takes is using `should` before we start the assertion chain.

I also piggy-back a small change that sets the changeSetId to not be
required on create for Entities. It will be eventually, but shouldn't be
yet. ;0

Love,
Adam