chore(cea): add `spawn_command_with_stdin` function
This change adds a new function to the `spawn_command` module called
`spawn_command_with_stdin` which will accept bytes to be fed to a
spawned process' stdin stream. The `spawn_command` function continues to
operate as before.

Additionally, several ergonomic updates were added to help when
consuming this library:

* Add the `si_cea::Entity` trait and `si_cea::CeaError` error type to
  the `si_cea::agent::dispath::prelude`
* Automatically prepare the `EntityEvent`'s `output_entity_event` field
  before calling the dispatch function, thus ensuring that the output
  field is always correctly set
* Add an `into_output` method on `CommandResult` to consume the `stdout`
  and `stderr` results when captured output was requested, thus saving
  needless cloning

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>