feat(kubernetes): add/impl `apply` action on `KubernetesDeployment`
This change adds an `apply` action to the `KubernetesDeploymentEntity`
which can be used for everything from creating a deployment to updating
it. The implementation leverages the `KubectlCommand` work prior work,
meaning that the `kubectl` program must be present on the process'
`$PATH` for this to work. Later work may graduate to use a client
library, another service/process, or another strategy to remove this
dependency from our stack.

Note: To temporarily see our system work, a default Kubernetes namespace
of `si` will be used on all `kubectl` calls. To override this default
you can set the `KUBERNETES_NAMESPACE` environment variable. While not
ideal for two reasons (i.e. hardcoded and unchangeable defaults == bad,
secret environment variables changing the behavior of library code ==
badder?), this isn't a forever code flow. Take note future team members:
Fletcher doesn't like this approach in general but is knowingly breaking
a rule ;)

Depends on #117

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>