feat(kubernetes): add `KubectlCommand` interface with k8s (#117)
This change adds a `KubectlCommand` type to the system which wraps an
async `Command` type. Currently the following methods are supported:

* `apply`: calls `kubectl apply` and takes its input on `stdin`
* `get_deployments`: lists all deployments in a namespace
* `get_deployment_last_applied`: calls `kubectl apply view-last-applied`
  to get the contents of the last applied template

This type heavily uses the `kubectl` command which must be present on
the prcesses `$PATH` and will return an error if the command cannot be
found.

Finally, to better directly try and test `KubectlCommand` 3 crate
examples are included to exercise these methods. The `kubectl-apply`
example will expect a deployment YAML to be provided on stdin so it can
be run with:

```console
$ cd components/si-kubernetes
$ cargo run --example kubectl-apply <nginx-deployment
```

Note that for the examples, a namespace of `si` will be used, which
should already be present.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>