chore(si-web-app): Switching to vuex and apollo directly
This adds an apollo API module, and wires it in to vuex. It then
replaces the 'auth' modules in utils, with a 'user' module loaded in the
store.

Unfortunately, my editor touched a lot of files in the process... so for
all the components, you're getting some lint cleanup that was
unintentional. The only edit that matters in them is the removal of the
(unused) `auth` import, which now references a file that doesn't exist.