feat: Change set execution and CRUD (#129)
This implements fully working Change Sets. When you create an Entity,
Update an entity, Delete an entity, or run an Action you are required to
pass a change set id. The results get stored as entries, which we then
walk when you call `changeSetExecute`.

This is a very janky implementation of Execute - it returns immediately,
and if you want to check if it is done, you have to poll. But! It is
async, and it does handle the dispatch use case for actions.

But it's a good starting place for when you don't have any test coverage
:)

This commit also includes the fixes to use changeSet's from the UI, and
refactors the UI to use the vuex store exclusively. Removing vue-apollo
will be a follow on commit.