feat(si-web-app): ChangeSet create and update with vuex (#144)
This allows for the creation and updating of new entities based on their
state in the vuex store. It is working, but has several known bugs:

1) If you update fields too quickly, particularly in a repeated object,
   you will wind up with only half your data saved. This is because the
   store itself is atomic, and the way I'm watching for edits isn't
   quite as effective as you would hope.

2) Numerous style and layout issues, in particular with how we want
   repeated objects to be rendered. Again, its working, but it will need
   love from Alex to feel good.

3) Speed. When updating maps or repeated objects, we currently have to
   round-trip all the way to the graphql backend before we render the
   updated data. This is a side-effect again of how the persistent store
   is working - I have an idea on how to fix this as well.

4) Occasionally, you will wind up with a partial load of an entity that
   doesn't refresh itself. This is because of the timing in which we
   call events, and the way vue detects that there are changes to be
   made.

I think this code is good enough to be merged, so that Alex can start
making things real, and we can start using it with multiple entity
types. It is going to need a refactor before we ship, but I think that
should happen after we have a chance to be certain this is the model we
want, rather than take any longer to ship this code.