feat(si-web-app): delete works
Delete works! You can now mark an entity as deleted, and when the
changeset executes, it will be deleted.

We have a bug in the flow where when you delete or create entities in a
changeset, they do not correctly update in the schematic panel until
after a refresh. Will fix that in another PR.