feat(core): add `Application` & `System` component & entity (#149)
This is a first pass at our Application and System concepts which are
fairly property and constraint-free at the moment. Right now they are
more important as objects in the system that can be referenced. We'll
see if these should be modeled as Components/Entities or something more
lightweight going forward.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>