feat: no more validation on request types (#152)
For now, we are disabling any kind of validation whatsoever on input
types. This will enable us to deal with entity creation and move on to
more interesting things, while we eventually circle back to both how we
deal with default values, picking components, validation, and more.

For now - enjoy your commented out overlord.