chore(transport): refactor MQTT code into a Transport abstraction
This change is rather large, but oddly should affect the external
behavior of the services and entire system in general.

In order to enforce consistent communication patterns between the core
system and semi-autonomous edge entities which interface with real world
resources--which we currently call Agents--a "Transport" abstraction is
introduced.

A `Transport` can send messages and also subscribe to a stream of
messages, implemented as a Rust `Stream`. Each message has two levels of
representation. The lowest levels of representation is a `WireMessage`
which encodes the target MQTT topic (or the source topic on
subscription), the quality of service (codified by a `QoS` type), an
optional response topic which can be used for RPC like patterns, a
payload type hint which is analogous to the `typeName` in the Registry,
and the message payload as raw bytes. The second representation is a
`Message<T>`, where `T` is the concrete Rust type corresponding to this
message. In this way, a `Message<T>` can be converted down to a
`WireMessage` and a `WireMessage` can be converted up into a
`Message<T>` (the `si-tranport` crate provides these conversions). Note
that the `ChangeSet` work dispatch, described below, uses a custom
deserializing into a `serde_json::Value` which is still compatible with
both `WireMessage`s and `Message<T>`s. In this way, the system retains
both "network packet" style transmission and static type safety where
required.

Individual Agents (formerly called `AgentServer`s) are responsible for
subscribing to topics of interest relating to their domain (now codified
as a `Topic`) and have type-aware static dispatch for each "object type"
(which is an analog to the Registry's `typeName` property). Each of
these "dispatchers" are configured and added to an `Agent` before the
Agent is set to run, thus giving us dynamically configurable dispatch at
runtime which should support future scenarios where an `Agent` is
deployed on its own in a customer's network or in another location on
"the edge" (i.e. out of our direct network of control). Each dispatcher
implementation knows how to dispatch a supported "action" to a handling
function, as implemented by human developers (that is, *not* code
generated). Each dispatched message has all the context necessary for
the Agent to perform its task, including a handle to a `Transport` which
can be used to stream back results and sent a "finalized" message. In
addition, this dispatched function is provided a "response topic" (now
codified as a `Header`) which can be used by the dispatched function to
supply status and other updates as necessary. Finally, as multiple
dispatchers can be added to an `Agent`, and are free from concrete
typing in a way that the Agent must be closely aware, a single Agent can
be responsible for multiple object types in one instance.

In light of the `ChangeSet` concept in our system, it turns out that a
`ChangeSet` is exclusively responsible for dispatching work to agents
and thus the formerly named `AgentClient` has been removed. In its
place, an `AgentExecuteSender` is provided in the `si-account` service
to pump work out to the Agents. As the account service is not concretely
aware of all types from other services, its work and data manipulation
is handled exclusively via the `serde_json` crate. However, since a
`Transport` message is serialized and transmitted with a `WireMessage`,
the payload remains as plain JSON and so this part of our system should
be able to continue with this strategy in a "safe" manner going forward.
The only caveat is if the `AgentExecuteSender` started to manipulate the
data type structure in a way that would corrupt a later deserialization
by other services, then we would see deserializing errors, but in the
domain service ultimately responsible for maintaining data and
persistence integrity.

Finally, the finalizing implementation was revisited to build on all the
learning and functionality above. Now called a `FinalizedListener`
(formerly called an `AgentFinalizer`), a single listener can handle
multiple object types in one instance. In addition to this, the actual
implementation of "finalizing" an object has been pushed into the object
itself, alongside its persistence implementation. Concretely, this means
that a `KubernetesDeployment` will have a `finalize(&mut self, db: &db)`
function, as does all other non-basic object types. Currently, only
`EntityEvent`s are finalized and have a special implementation that is
aware of their associated `Entity`, so all `EntityEvent`s types have
this implementation with all other object types having a simple "save"
implementation. The end result is that any object type can be finalized.
Additionally, as there is now no special code required to finalize an
arbitrary object, a full implementation for each object type's
"finalizer" is fully code generated and available for consumption by a
service author.

This ended up not being small nor fast, although it's my hope that the
timing on further locking this down was about right and this leaves us
with a very operable solution when dealing with remote Agents.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>