fix(transport): fix `Display` impl for `AgentData` header & topic (#166)
An agent data topic is a "data" topic, rather than a "command" topic,
where a data topic subscription would be something like:

```
$shared/dt/agent/+/+/+/+/+/+/+/application_entity_event/+/finalize
```

vs a command topic subscription which looks like:

```
$shared/cmd/agent/+/+/+/+/+/+/+/application_entity_event/+/execute
```

It turns out that `AgentDataHeader` & `AgentDataTopic` had an incorrect
formatting, where `cmd` was being inserted, where `dt` was expected. The
complementary `FromStr` impl was correctly parsing the data topics and
headers and therefore would error out if a data header or topic was
serialized to a `String`, then deserialized back.

Interestingly, this was discovered when booting `si-core` and
`si-kubernetes`, both of which spawn a `FinalizedListener`, which
subscribe to data topics (rather than command topics). The code that
spawns this subsystem would unconditionally fail on setup, which I
suspect lead to an honest-to-god segfault/core dump. I'm still a bit
baffled as to why, as the setup code should be taking place in the main
thread of execution, before `tokio::spawn(listener.run())` is run,
but...it's gone now?

Anyway, this was a bug regardless and a good lesson to add unit tests to
this part of our codebase. I ignored my instincts here in the interests
of finishing this work out and it may make good sense to add a few tests
in the next few weeks to shore up the basic behaviors in si-transport.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>