chore(vernemq): update VerneMQ container to official upstream
This change swaps our VerneMQ image to `vernemq/vernemq` which appears
to be actively maintained by the upstream company and project. This also
includes tags for versioned releases, and so the deployed version of
VerneMQ is currently controlled via a `VERSION` variable in the `run.sh`
script.

Most importantly, the MQTT v5 protocol is explicitly enabled, as we're
starting to use features from this protocol version.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>