fix(transport): use a `$share` topic, not a `$shared` topic
This is a straight up bug fix where a "shared" topic subscription should
be prefixed with `$share/`, but in referring to shared topics, the
literal `$shared/` topic prefix was added instead. And I think that's
probably the better choice, but that choice is not up to me, so I bow to
smarter people than myself.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>