feat: allow for debugging events
This commit allows for (basic) debugging of events. The Event Bar now
has clickable chevrons for each line, which when expanded show more
information about the event.

It also now includes an event for when an EntityEvent is finalized, so
we can get the log messages. Its messy and wasteful, but its going to
work.

It doesn't stream the event logs as the event is firing - this should be
fine for now, as the events we are firing just don't take very long.
Instead, it shows you the final results, with the output inline.