bug: fix the save loop
This fixes the save flow [ch453], at least for now. Maps still have a
bug with their data being lost on entry, and if you don't want for the
save, you can loose some data on adding to a repeated object.

I have plans to fix both bugs, but it means a bigger refactor of the
whole flow, which really needs a new service (which is also on the
board).

For now, you should be good to edit fields as fast as you can type.