bug: persist the store locally for faster loading
This is a partial solution - it stores the data in the browsers
IndexedDB every time the Vuex store is mutated. The result is that we
get the full application state all the time, no matter what.

That means reloading is really fast, because all your application state
is stored on disk. Yay!

The bad part: it will only work for a single user - if mutations happen
outside your web browser, all bets are off.

The medium part: I added a "refresh" logout button down near the user
icon. If you click it, the indexeddb will be dropped, and you will be
logged out. If you then shift-refresh the browser, you will be starting
again from scratch.

Eventually, this behavior needs to be completely thought out -
essentially, vuex is great (reactive state! central!) and also not so
great (double persisting to the API!). I think it's all fine for now,
but as soon as we have multiple users, we will want to extend this
functionality.

Love,
Adam