feat: Resources visualization (#199)
This PR provides resource visualization. It allows for the creation and
updating of resources, which track their status and health. A node can
have many resources. Right this moment, they all have only one. ;)

When you hit the `sync` button, we flip all the resources to unknown,
and then after an unspecified period of time flip them all green again.

* Users can see the list of resources, with circles for status. [ch550]
* Resource tooltips provide details when you hover over the circle.
  [ch549]
* Turn off telemetry, since the otel collector is a pain in my ass.
* Relocate the system drop down menu [ch553]
* Ability to sync resources from a button [ch552]