chore: pin to Rust nightly 2020-09-27 (#203)
Using the current Rust release 1.46.0, the following compilation error
occurs in the `si-core` crate:

```
error: reached the type-length limit while instantiating `std::thread::LocalKey::<std::cel...ult::Result<(), anyhow::Error>>>`
   --> /home/fnichol/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/src/libstd/thread/local.rs:235:5
    |
235 | /     pub fn with<F, R>(&'static self, f: F) -> R
236 | |     where
237 | |         F: FnOnce(&T) -> R,
238 | |     {
...   |
242 | |         )
243 | |     }
    | |_____^
    |
    = note: consider adding a `#![type_length_limit="1134387"]` attribute to your crate

error: aborting due to previous error

error: could not compile `si-core`.
```

This seems to be related to a regression which was introduced between
the last stable release and current, as mentioned in
rust-lang/rust#75992. As there is no fix landed yet in stable, and when
trying to follow the compiler hint leads to a longer compile time and a
suggestion to further increase the value for `type_length_limit`, the
easiest course of action is for the moment to pin our work to Rust
`1.45.2`. This fix pins us to a nightly version that seems to resolve
the issue.

Honestly, this is a first for me where what appears to be a Rust
compiler issue caused a compiletime regression in stable. Likely this
doesn't stick around for long.

References: https://github.com/rust-lang/rust/issues/75992

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>