feat: services visualization and deployment button
This PR fixes [ch548], [ch545], and [ch546].

It adds a square for each service in the application, with that services
health and status (which will eventually be a roll-up of the
status/health of all objects underneath it). When you hover over them,
you get a helpful tooltip.

When you are in a changeSet, you can now hit the "deploy" button, and it
will automatically schedule a deploy of the entire application (which will recursively
call the services, etc) in the change set, and then execute it.

The results are a closed changeset and a deployed thingamajig.

We're into it.

Love,
Adam and Fletcher