chore: docker containers and web app build
This PR sets up simpler docker containers, and fixes the web app build
to work with netlify.

It makes the CI actions work again, with testing for si-sdf. It does not
include the (as yet) un-rebuild si-pedant integration testing, which
we absolutely should do, but need to think a little differently about
how (and when) we want to run them.

This fixes [ch657].

* Docker images for si-veritech, si-sdf, couchbase, nats, and otelcol.
* Configured netlify to deploy si-web-app.
* Improved the CI pipeline, so that it runs much quicker.